Switched up routes and app rendering. Set default route to Productcontainer for now. Should be switched to productscontainer  when ready.
