+++ app/routes.jsx
@@ -1,0 +1,39 @@
+'use strict'
+import React from 'react'
+import {Router, Route, IndexRedirect, browserHistory} from 'react-router'
+import {render} from 'react-dom'
+import {connect, Provider} from 'react-redux'
+
+import store from './store'
+import Jokes from './components/Jokes'
+import Login from './components/Login'
+import WhoAmI from './components/WhoAmI'
+import App from './components/App'
+import ProductContainer from './containers/ProductContainer'
+
+const ExampleApp = connect(
+  ({ auth }) => ({ user: auth })
+) (
+  ({ user, children }) =>
+    <div>
+      <nav>
+        {user ? <WhoAmI/> : <Login/>}
+      </nav>
+      {children}
+    </div>
+)
+
+
+
+export default function Root () {
+  return (
+    <Provider store={store}>
+      <Router history={browserHistory}>
+        <Route path="/" component={App} >
+          <Route path="/product" component={ProductContainer}/>
+          <IndexRedirect to="/product"/>
+        </Route>
+      </Router>
+    </Provider>
+  );
+}
