+++ public/index.html
@@ -1,10 +1,10 @@
-<!doctype html>
+<!DOCTYPE html>
 <html>
   <head>
     <link href="/bootstrap/css/bootstrap.min.css" rel="stylesheet">
   </head>
   <body>
-    <div id=main></div>
+    <div id='app'></div>
     <script async defer src=/bundle.js></script>
     <script src="/bootstrap/js/bootstrap.min.js"></script>
   </body>
